backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiohttp==3.9.1
openai==1.10.0
pydantic==2.5.3
pydantic[email]==2.5.3
firebase-admin==6.4.0